Move integration tests to root `test/` directory
