+++ test/integration/twitter-adapter-playwright/UserProfileViewAdapter.test.ts
@@ -1,6 +1,6 @@
 import { Browser, BrowserContext, chromium, Page } from "playwright";
-import { UserProfileView } from "../twitter-api/ports/UserProfileView";
-import { UserProfileViewAdapter } from "./UserProfileViewAdapter";
+import { UserProfileView } from "../../../src/twitter-api/ports/UserProfileView";
+import { UserProfileViewAdapter } from "../../../src/twitter-adapter-playwright/UserProfileViewAdapter";
 
 jest.setTimeout(30000);
